+++ src/renderer/src/components/github-project/project-view-wrapper-source-context-boundary.test.ts
@@ -1,5 +1,3 @@
-// @vitest-environment happy-dom
-
 import { readFileSync } from 'node:fs'
 import { join } from 'node:path'
 import { describe, expect, it } from 'vitest'
@@ -21,7 +19,7 @@ function sourceBetween(source: string, startPattern: string, endPattern: string)
 
 describe('ProjectViewWrapper GitHub source context boundary', () => {
   it('builds project work items with a host-pinned repository identity', async () => {
-    const { buildProjectWorkItem } = await import('./ProjectViewWrapper')
+    const { buildProjectWorkItem } = await import('./project-work-item')
     const row: GitHubProjectRow = {
       id: 'PVTI_1',
       itemType: 'PULL_REQUEST',
